+++ src/core/player-view.ts
@@ -6,6 +6,8 @@
  * https://opensource.org/licenses/MIT.
  */
 
+import { Ctx, PlayerID } from '../types';
+
 /**
  * PlayerView reducers.
  */
@@ -17,7 +19,7 @@ export const PlayerView = {
    * removes all the keys in `players`, except for the one
    * corresponding to the current playerID.
    */
-  STRIP_SECRETS: (G, ctx, playerID) => {
+  STRIP_SECRETS: (G: any, ctx: Ctx, playerID: PlayerID) => {
     let r = { ...G };
 
     if (r.secret !== undefined) {